refactor(services): status-service를 provider-registry 기반으로 분리 (#24)
기존 388줄짜리 status-service.js를 다음 4개 파일로 분해한다.

- services/codex-provider.js (146 LOC)
  getCodexSnapshot, selectCodexAuthSource, filterRealCodexAccounts
  및 resolveCodexProfiles/getAgentStoreProfiles/mapAccountToProfile/
  createCodexFailureSnapshot(private)
- services/claude-provider.js (190 LOC)
  getClaudeSnapshot, buildClaudeSnapshot, resolveClaudeProfileFromSnapshot,
  selectClaudeAuthSource 및 loadAgentStoreClaudeAccounts/
  createClaudeNetworkFailureSnapshot(private)
- services/provider-registry.js (32 LOC)
  { id, getSnapshot } spec 배열 + runProviderSnapshots 헬퍼
- services/status-service.js (49 LOC)
  config 로드 + registry 순회. provider별 helper는 re-export로 backward
  compat 유지.

새 provider를 추가할 때는 provider-registry.js의 PROVIDER_REGISTRY
배열에 한 줄을 더하면 된다. 기존 테스트 import 경로는 그대로라 회귀 없음.

247 tests pass. CLI end-to-end도 동일 출력 확인.

closes #24
